(step4) 코멘트반영
